Support resource metering for other os
