+++ components/resource_metering/Cargo.toml
@@ -13,6 +13,7 @@ prost-codec = [
   "kvproto/prost-codec",
   "grpcio/prost-codec",
 ]
+failpoints = ["fail/failpoints"]
 
 [dependencies]
 lazy_static = "1.3"
@@ -33,9 +34,15 @@ slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debu
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "d592f88e4dbba5eb439998463054f1a44fbf17b9" }
 futures = "0.3"
 pdqselect = "0.1"
+fail = "0.4"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 procinfo = { git = "https://github.com/tikv/procinfo-rs", rev = "5125fc1a69496b73b26b3c08b6e8afc3c665a56e" }
 
 [dev-dependencies]
 rand = "0.8"
+
+[[test]]
+name = "test-cpu-recorder-linux"
+path = "tests/cpu_recorder_linux.rs"
+required-features = ["failpoints"]
