--- components/resource_metering/src/cpu/collector/dummy.rs
@@ -1,8 +1,0 @@
-// Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
-
-use crate::cpu::collector::Collector;
-
-pub struct CollectorHandle;
-pub fn register_collector(_collector: Box<dyn Collector>) -> CollectorHandle {
-    CollectorHandle
-}
