+++ components/resource_metering/src/cpu/collector/mod.rs
@@ -1,20 +1,13 @@
 // Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::cpu::recorder::CpuRecords;
+use crate::cpu::recorder::records::CpuRecords;
 
+use std::sync::atomic::AtomicU64;
+use std::sync::atomic::Ordering::Relaxed;
 use std::sync::Arc;
 
-#[cfg(target_os = "linux")]
-mod linux;
-#[cfg(target_os = "linux")]
-pub use linux::{register_collector, CollectorHandle};
-#[cfg(target_os = "linux")]
-pub(crate) use linux::{CollectorRegistrationMsg, COLLECTOR_REGISTRATION_CHANNEL};
-
-#[cfg(not(target_os = "linux"))]
-mod dummy;
-#[cfg(not(target_os = "linux"))]
-pub use dummy::{register_collector, CollectorHandle};
+use crossbeam::channel::{unbounded, Receiver, Sender};
+use lazy_static::lazy_static;
 
 pub trait Collector: Send {
     fn collect(&self, records: Arc<CpuRecords>);
@@ -22,3 +15,46 @@ pub trait Collector: Send {
 
 #[derive(Copy, Clone, Default, Debug, Eq, PartialEq, Hash)]
 pub struct CollectorId(pub(crate) u64);
+
+lazy_static! {
+    pub(crate) static ref COLLECTOR_REGISTRATION_CHANNEL: (
+        Sender<CollectorRegistrationMsg>,
+        Receiver<CollectorRegistrationMsg>
+    ) = unbounded();
+}
+
+pub(crate) enum CollectorRegistrationMsg {
+    Register {
+        collector: Box<dyn Collector>,
+        id: CollectorId,
+    },
+    Unregister {
+        id: CollectorId,
+    },
+}
+
+pub fn register_collector(collector: Box<dyn Collector>) -> CollectorHandle {
+    lazy_static! {
+        static ref NEXT_COLLECTOR_ID: Arc<AtomicU64> = Arc::new(AtomicU64::new(1));
+    }
+
+    let id = CollectorId(NEXT_COLLECTOR_ID.fetch_add(1, Relaxed));
+    COLLECTOR_REGISTRATION_CHANNEL
+        .0
+        .send(CollectorRegistrationMsg::Register { collector, id })
+        .ok();
+    CollectorHandle { id }
+}
+
+pub struct CollectorHandle {
+    id: CollectorId,
+}
+
+impl Drop for CollectorHandle {
+    fn drop(&mut self) {
+        COLLECTOR_REGISTRATION_CHANNEL
+            .0
+            .send(CollectorRegistrationMsg::Unregister { id: self.id })
+            .ok();
+    }
+}
