+++ components/resource_metering/tests/mod.rs
@@ -1,5 +1,4 @@
 // Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
 
-pub mod collector;
-pub mod recorder;
-pub mod thread;
+#[cfg(target_os = "linux")]
+mod cpu_recorder_linux;
