+++ src/coprocessor/endpoint.rs
@@ -33,7 +33,8 @@ use crate::coprocessor::tracker::Tracker;
 use crate::coprocessor::*;
 use concurrency_manager::ConcurrencyManager;
 use engine_rocks::PerfLevel;
-use resource_metering::{cpu::FutureExt, ResourceMeteringTag};
+use resource_metering::future_ext::FutureExt;
+use resource_metering::tag::ResourceMeteringTag;
 use txn_types::Lock;
 
 /// Requests that need time of less than `LIGHT_TASK_THRESHOLD` is considered as light ones,
