+++ tests/integrations/config/dynamic/resource_metering.rs
@@ -1,61 +1,56 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use std::sync::mpsc::{channel, Receiver, Sender};
+use std::sync::{Arc, Mutex};
 
-use resource_metering::cpu::recorder::RecorderHandle;
-use resource_metering::reporter::Task;
-use resource_metering::ConfigManager;
+use resource_metering::ConfigManagerBuilder;
 use tikv::config::{ConfigController, Module, TiKvConfig};
 use tikv_util::config::ReadableDuration;
-use tikv_util::worker::{LazyWorker, Runnable};
-
-pub struct MockResourceMeteringReporter {
-    tx: Sender<Task>,
-}
-
-impl MockResourceMeteringReporter {
-    fn new(tx: Sender<Task>) -> Self {
-        MockResourceMeteringReporter { tx }
-    }
-}
-
-impl Runnable for MockResourceMeteringReporter {
-    type Task = Task;
-
-    fn run(&mut self, task: Self::Task) {
-        self.tx.send(task).unwrap();
-    }
-}
-
-fn setup_cfg_manager(
-    config: TiKvConfig,
-) -> (ConfigController, Receiver<Task>, Box<LazyWorker<Task>>) {
-    let mut worker = Box::new(LazyWorker::new("resource-metering-reporter"));
-    let scheduler = worker.scheduler();
-
-    let (tx, rx) = channel();
-
-    let resource_metering_config = config.resource_metering.clone();
-    let cfg_controller = ConfigController::new(config);
-    cfg_controller.register(
-        Module::ResourceMetering,
-        Box::new(ConfigManager::new(
-            resource_metering_config,
-            scheduler,
-            RecorderHandle::default(),
-        )),
-    );
-
-    worker.start(MockResourceMeteringReporter::new(tx));
-    (cfg_controller, rx, worker)
-}
 
 #[test]
 fn test_update_resource_metering_agent_config() {
     let (mut config, _dir) = TiKvConfig::with_tmp().unwrap();
     config.validate().unwrap();
 
-    let (cfg_controller, rx, worker) = setup_cfg_manager(config);
+    let cfg_controller = ConfigController::new(config.clone());
+
+    let config = Arc::new(Mutex::new(config.resource_metering.clone()));
+    let cfg_mgr = {
+        let mut cfg_mgr_builder = ConfigManagerBuilder::new();
+
+        let cfg = config.clone();
+        cfg_mgr_builder.on_change_max_resource_groups(move |m| {
+            let mut c = cfg.lock().unwrap();
+            c.max_resource_groups = m;
+        });
+
+        let cfg = config.clone();
+        cfg_mgr_builder.on_change_agent_address(move |a| {
+            let mut c = cfg.lock().unwrap();
+            c.agent_address = a.to_owned();
+        });
+
+        let cfg = config.clone();
+        cfg_mgr_builder.on_change_enabled(move |e| {
+            let mut c = cfg.lock().unwrap();
+            c.enabled = e;
+        });
+
+        let cfg = config.clone();
+        cfg_mgr_builder.on_change_precision(move |p| {
+            let mut c = cfg.lock().unwrap();
+            c.precision = p;
+        });
+
+        let cfg = config.clone();
+        cfg_mgr_builder.on_change_report_agent_interval(move |r| {
+            let mut c = cfg.lock().unwrap();
+            c.report_agent_interval = r;
+        });
+
+        cfg_mgr_builder.build()
+    };
+
+    cfg_controller.register(Module::ResourceMetering, Box::new(cfg_mgr));
 
     let change = {
         let mut m = std::collections::HashMap::new();
@@ -84,14 +79,5 @@ fn test_update_resource_metering_agent_config() {
     assert_eq!(new_config.report_agent_interval, ReadableDuration::secs(80));
     assert_eq!(new_config.max_resource_groups, 3000);
 
-    let task = rx.recv().unwrap();
-
-    match task {
-        Task::ConfigChange(config) => {
-            assert_eq!(config, new_config)
-        }
-        _ => unreachable!(),
-    }
-
-    worker.stop_worker();
+    assert_eq!(&*config.lock().unwrap(), &new_config);
 }
